fix: /index.html 요청에 대한 처리 로직을 플러그인으로 분리
